strategies/insider/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Data providers
yfinance>=0.2.32

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Optional: For EDGAR XML parsing (built-in)
# lxml>=4.9.0

# Optional: For testing
pytest>=7.4.0
pytest-cov>=4.1.0
